Extracing the actual serializer code
serializer.js using to do to things:
1. Interpret code (global code, and possibly speculatively initializing more code).
2. Serialize the heap. Actually, twice, as we do two passes: One to count if values are used more than once (which means we need to allocate a name for it), and one more time to generate the final code.

This mechanical refactoring pulls out the actual serialization. As a side effect, we no longer this funny re-initialization between passes.

The serialization functionality of ResidualHeapSerializer now matches the ResidualHeapVisitor closely. The serializer is just doing the extra required work of actually producing code, plus trying hard to generate efficient code...

The names "serializer" vs "ResidualHeapSerializer" are not great. The "serializer" should probably be renamed next, as it does more than just serialization: It's really implementing the prepack-functionality of interpreting+serializing. In any case, the next rename should come in a separate pull requests to keep this one simple.
